+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/Clothes/EffectEquipFurryEars.lua
@@ -1,0 +1,16 @@
+EffectEquipFurryEars = ChaosEffectBase:derive("EffectEquipFurryEars", "equip_furry_ears")
+
+function EffectEquipFurryEars:OnStart()
+    ChaosEffectBase:OnStart()
+    local player = getPlayer()
+    if not player then return end
+
+    local inventory = player:getInventory()
+    if not inventory then return end
+
+    local item = inventory:AddItem("Base.Hat_FurryEars")
+    if item then
+        ChaosPlayer.EquipClothes(player, item)
+        ChaosPlayer.SayLineNewItem(player, item)
+    end
+end